refactoring and adding rule update
